User Authenticate

+++ controller/expenseController.js
@@ -1,4 +1,5 @@
 const Expense = require("../model/expensListTable");
+const User = require("../model/signupTable");
 
 
 //ADD expense
@@ -7,7 +8,7 @@ const addExpense= async(req,res,next)=>{
     const{amount,description,category}=req.body
 
     const data= await Expense.create(
-        {amount,description,category});
+        {amount,description,category,userId:req.user.id});
     res.status(201).json({newExpenseDetail:data});
     } catch (err) {
         console.log(err);
@@ -21,10 +22,12 @@ const addExpense= async(req,res,next)=>{
 const getExpense=async(req,res,next)=>{
 
     try {
-        const expenses= await Expense.findAll();
-    res.status(200).json({allExpenses:expenses});
+        await Expense.findAll({where:{userId:req.user.id}}).then(expenses=>{
+            return res.status(200).json({allExpenses:expenses});
+        })
+            
     } catch (error) {
-        console.log(err);
+        //console.log(err);
         res.status(500).json(error);
     }
 
@@ -32,7 +35,7 @@ const getExpense=async(req,res,next)=>{
 //Delete Expense
 
 const deleteExpense=async(req,res)=>{
-    console.log("113");
+    //console.log("113");
     try {
         
         if(req.params.id=='undefined'){
@@ -40,8 +43,13 @@ const deleteExpense=async(req,res)=>{
             return res.status(404).json({err:'id is missing'})
         }
         const eId=req.params.id;
-        await Expense.destroy({where:{id:eId}});
-        res.sendStatus(200);
+        await Expense.destroy({where:{id:eId,userId:req.user.id}}).then(data=>{
+            if(data===0){
+                return res.status(404).json({message:"Expense not belong to user"})
+            } else{
+                res.status(200).json({message:"Deleted Successfully"});
+            }
+            })
     } catch (err) {
         console.log(err);
         res.status(500).json(err);
